Anki Note Generator From YT Video/pyproject.toml
[project]
name = "anki-note-generator-from-yt-video"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "genanki>=0.13.1",
    "groq>=0.31.0",
    "google-genai>=1.30.0",
    "questionary>=2.1.0",
    "yt-dlp>=2025.8.11",
    "aiohttp>=3.9.5",
    "pydantic>=2.8.2",
    "youtube-transcript-api>=0.6.2",
    "json5>=0.9.25",
]

[dependency-groups]
dev = [
    "pytest>=8.4.1",
    "pytest-asyncio>=0.23.8",
]

[tool.ruff.lint]
per-file-ignores = { "tests/*.py" = ["I001"] }